+++ src/simulation/simulation.jl
@@ -1,0 +1,13 @@
+__precompile__(true)
+module Simulation
+
+using Reexport
+
+# Simulation submodules
+include("integrators.jl")
+include("propagators.jl")
+
+@reexport using SatelliteDynamics.Simulation.Integrators
+@reexport using SatelliteDynamics.Simulation.Propagators
+
+end # Simulation Module